pyproject.toml
[project]
name = "bilibili-analyzer"
version = "0.1.0"
description = "Bilibili视频分析系统"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
dependencies = [
    "flask>=2.3.0",
    "flask-sqlalchemy>=3.0.0",
    "flask-migrate>=4.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "apscheduler>=3.10.0",
    "celery>=5.3.0",
    "redis>=4.6.0",
]
requires-python = ">=3.8"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm.dev-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-flask>=1.3.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
    "coverage>=7.3.0",
]

[tool.pdm.scripts]
start = "python app.py"
test = "pytest"
lint = "flake8 bilibili_analyzer/"
format = "black bilibili_analyzer/ app.py config.py"
type-check = "mypy bilibili_analyzer/"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true